23 sprint

+++ app/src/main/java/com/example/playlistmaker/media/data/db/PlaylistEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "playlist_table")
 class PlaylistEntity(
     @PrimaryKey(autoGenerate = true)
-    val playlistId: Int,
+    val playlistId: Long,
     val playlistName: String,
     val playlistDescription: String?,
     val coverPath: String?,
